refactor: move ModelSelect component to separate file
